refactor(EST-3998): split version conversions into per-version modules
Address review feedback: convertions.py becomes a package with one
file per version conversion (version_1_to_2.py, version_2_to_3.py),
restore the original v1_to_v2 docstring, and trim the v2_to_v3
docstring to drop historical reasoning.
