+++ src/django_app/tables/import_export/version_conversions/convertions/__init__.py
@@ -1,0 +1,11 @@
+# ruff: noqa: F401
+from tables.import_export.version_conversions.convertions import (
+    version_1_to_2,
+    version_2_to_3,
+)
+from tables.import_export.version_conversions.convertions.version_1_to_2 import (
+    v1_to_v2,
+)
+from tables.import_export.version_conversions.convertions.version_2_to_3 import (
+    v2_to_v3,
+)
